+++ R/nse-force.R
@@ -84,25 +84,40 @@
 #'
 #' list2(!!name := 1 + 2)
 #' exprs(!!name := 1 + 2)
-#' quos(!!name := 1 + 2)
 #' ```
 #'
 #' Like `=`, the `:=` operator expects strings or symbols on its LHS.
+#'
+#' Since unquoting names is related to interpolating within a string
+#' with the glue package, we have made the glue syntax available on
+#' the LHS of `:=`:
+#'
+#' ```
+#' list2("{name}" := 1)
+#' tibble("{name}" := 1)
+#' ```
+#'
+#' You can also interpolate defused function arguments with double
+#' braces `{{`, similar to the curly-curly syntax:
+#'
+#' ```
+#' wrapper <- function(data, var) {
+#'   data %>% mutate("{{ var }}_foo" := {{ var }} * 2)
+#' }
+#' ```
 #'
 #' Currently, forcing names with `:=` only works in top level
 #' expressions. These are all valid:
 #'
 #' ```
-#' exprs(!!nm := x)
-#' tibble(!!nm := x)
-#' list2(!!nm := x)
+#' exprs("{name}" := x)
+#' tibble("{name}" := x)
 #' ```
 #'
 #' But deep-forcing names isn't supported:
 #'
 #' ```
-#' expr(this(is(deep(!!nm := x))))
-#' exprs(this(is(deep(!!nm := x))))
+#' exprs(this(is(deep("{name}" := x))))
 #' ```
 #'
 #'
@@ -349,4 +364,23 @@ expr_interp <- function(x, env = NULL) {
     x <- .Call(rlang_interp, x, env %||% parent.frame())
   }
   x
+}
+
+
+glue_unquote <- function(text, env = caller_env()) {
+  glue::glue(glue_first_pass(text, env = env), .envir = env)
+}
+glue_first_pass <- function(text, env = caller_env()) {
+  glue::glue(
+    text,
+    .open = "{{",
+    .close = "}}",
+    .transformer = glue_first_pass_eval,
+    .envir = env
+  )
+}
+glue_first_pass_eval <- function(text, env) {
+  text_expr <- parse_expr(text)
+  defused_expr <- eval_bare(call2(enexpr, text_expr), env)
+  as_label(defused_expr)
 }