+++ R/rlang.R
@@ -5,19 +5,21 @@ NULL
 is_same_body <- NULL
 
 
+on_package_load <- function(pkg, expr) {
+  if (isNamespaceLoaded(pkg)) {
+    expr
+  } else {
+    thunk <- function(...) expr
+    setHook(packageEvent(pkg, "onLoad"), thunk)
+  }
+}
+
+
 downstream_deps <- list(
   dplyr = c(min = "0.8.0", from = "0.4.0")
 )
 
 check_downstream_dep <- function(dep, pkg) {
-  if (!isNamespaceLoaded(pkg)) {
-    setHook(
-      packageEvent(pkg, "onLoad"),
-      function(...) check_downstream_dep(dep, pkg)
-    )
-    return()
-  }
-
   min <- dep[["min"]]
   from <- dep[["from"]]
   stopifnot(
@@ -54,6 +56,7 @@ check_downstream_dep <- function(dep, pkg) {
   warn(msg)
 }
 
+
 base_ns_env <- NULL
 base_pkg_env <- NULL
 
@@ -64,13 +67,18 @@ base_pkg_env <- NULL
     is_same_body <<- is_reference
   }
 
+  on_package_load("glue", .Call(rlang_glue_is_there))
+
   .Call(r_init_library)
-  .Call(rlang_library_load)
+  .Call(rlang_library_load, ns_env("rlang"))
 
   s3_register("pillar::pillar_shaft", "quosures", pillar_shaft.quosures)
   s3_register("pillar::type_sum", "quosures", type_sum.quosures)
 
-  map2(downstream_deps, names(downstream_deps), check_downstream_dep)
+  map2(downstream_deps, names(downstream_deps), function(dep, pkg) {
+    force(dep)
+    on_package_load(pkg, check_downstream_dep(dep, pkg))
+  })
 
   base_ns_env <<- ns_env("base")
   base_pkg_env <<- baseenv()